Removed some WIP duplication
