+++ lib/middleware/routes-metadata/routes-metadata.js
@@ -39,11 +39,7 @@ const {
   setControlNames,
   setRepeatable,
   setMultipartForm,
-  setMOJMultipartForm,
   processUploads,
-  processMOJUploads,
-  setUploads,
-  setMOJUploads,
   removeItem,
   addItem,
   processInput,
@@ -57,6 +53,12 @@ const {
   skipComponents,
   kludgeUpdates
 } = require('~/fb-runner-node/page/page')
+
+const setUploads = require('~/fb-runner-node/page/set-uploads/set-uploads')
+const setUploadControlsMaxSize = require('~/fb-runner-node/page/set-uploads/set-upload-controls-max-size')
+
+const setMOJUploads = require('~/fb-runner-node/page/set-moj-uploads/set-moj-uploads')
+const setMOJUploadControlsMaxSize = require('~/fb-runner-node/page/set-moj-uploads/set-moj-upload-controls-max-size')
 
 const {
   getInstanceController,
@@ -253,27 +255,31 @@ const pageHandler = async (req, res) => {
         components = []
       } = pageInstance
 
-      if (components.some(({_type}) => 'fileupload')) {
-        // Handle multi-part forms and uploads
+      const hasFUpload = components.some(({_type}) => 'fileupload')
+      const hasMUpload = components.some(({_type}) => 'mojUpload')
+
+      if (hasFUpload || hasMUpload) {
         pageInstance = await setMultipartForm(pageInstance, userData)
         if (pageInstance.encType) {
+          if (hasFUpload) {
+            setUploadControlsMaxSize(pageInstance)
+          }
+
+          if (hasMUpload) {
+            setMOJUploadControlsMaxSize(pageInstance)
+          }
+
           if (POST) {
             pageInstance = await processUploads(pageInstance, userData)
           }
 
-          pageInstance = await setUploads(pageInstance, userData)
-        }
-      }
-
-      if (components.some(({_type}) => 'mojUpload')) {
-        // Handle MOJ multi-part forms and uploads
-        pageInstance = await setMOJMultipartForm(pageInstance, userData)
-        if (pageInstance.encType) {
-          if (POST) {
-            pageInstance = await processMOJUploads(pageInstance, userData)
+          if (hasFUpload) {
+            pageInstance = await setUploads(pageInstance, userData)
           }
 
-          pageInstance = await setMOJUploads(pageInstance, userData)
+          if (hasMUpload) {
+            pageInstance = await setMOJUploads(pageInstance, userData)
+          }
         }
       }
     }
