+++ lib/page/page.js
@@ -5,10 +5,6 @@ const setControlNames = require('./set-control-names/set-control-names')
 const setRepeatable = require('./set-repeatable/set-repeatable')
 const setMultipartForm = require('./set-multipart-form/set-multipart-form')
 const processUploads = require('./process-uploads/process-uploads')
-const setUploads = require('./set-uploads/set-uploads')
-const setMOJMultipartForm = require('./set-moj-multipart-form/set-moj-multipart-form')
-const processMOJUploads = require('./process-moj-uploads/process-moj-uploads')
-const setMOJUploads = require('./set-moj-uploads/set-moj-uploads')
 const removeItem = require('./remove-item/remove-item')
 const addItem = require('./add-item/add-item')
 const processInput = require('./process-input/process-input')
@@ -30,10 +26,6 @@ module.exports = {
   setRepeatable,
   setMultipartForm,
   processUploads,
-  setUploads,
-  setMOJMultipartForm,
-  processMOJUploads,
-  setMOJUploads,
   removeItem,
   addItem,
   processInput,
