+++ lib/page/page.unit.spec.js
@@ -8,11 +8,7 @@ const expectedMethods = [
   'setControlNames',
   'setRepeatable',
   'setMultipartForm',
-  'setMOJMultipartForm',
   'processUploads',
-  'processMOJUploads',
-  'setUploads',
-  'setMOJUploads',
   'removeItem',
   'addItem',
   'processInput',
