--- lib/page/process-moj-uploads/get-allowed-moj-upload-controls.js
@@ -1,25 +1,0 @@
-require('@ministryofjustice/module-alias/register-module')(module)
-
-const {getUploadMaxFiles, getUploadFileCount} = require('~/fb-runner-node/page/utils/utils-uploads')
-
-const getAllowedMOJUploadControls = (userData, uploadControls) => {
-  const alloweduploadControls = []
-  uploadControls.forEach(uploadControl => {
-    const uploadName = uploadControl.name
-    const maxFiles = getUploadMaxFiles(uploadControl)
-    const currentFiles = getUploadFileCount(uploadControl, userData)
-    let allowable = maxFiles - currentFiles
-    let uploadCount = 1
-    while (allowable > 0) {
-      alloweduploadControls.push({
-        name: `${uploadName}[${uploadCount}]`,
-        maxCount: 1
-      })
-      uploadCount++
-      allowable--
-    }
-  })
-  return alloweduploadControls
-}
-
-module.exports = getAllowedMOJUploadControls
