--- lib/page/process-moj-uploads/multer/lib/counter.js
@@ -1,2 +1,0 @@
-const Counter = require('multer/lib/counter')
-module.exports = Counter
