--- lib/page/process-moj-uploads/multer/lib/file-appender.js
@@ -1,2 +1,0 @@
-const FileAppender = require('multer/lib/file-appender')
-module.exports = FileAppender
