--- lib/page/process-moj-uploads/multer/lib/multer-error.js
@@ -1,2 +1,0 @@
-const MulterError = require('multer/lib/multer-error')
-module.exports = MulterError
