--- lib/page/process-moj-uploads/multer/lib/remove-uploaded-files.js
@@ -1,2 +1,0 @@
-const removeUploadedFiles = require('multer/lib/remove-uploaded-files')
-module.exports = removeUploadedFiles
