+++ lib/page/process-uploads/process-uploads.js
@@ -74,8 +74,6 @@ module.exports = async function processUploads (pageInstance, userData) {
 
     const endUploadControlsMetrics = uploadControlsMetrics.startTimer()
     const uploadControls = getUploadControls(pageInstance)
-    // .filter(({_type}) => _type === 'fileupload')
-
     endUploadControlsMetrics()
 
     const endAllowedUploadControlsMetrics = allowedUploadControlsMetrics.startTimer()
