+++ lib/page/set-multipart-form/set-multipart-form.js
@@ -1,7 +1,6 @@
 require('@ministryofjustice/module-alias/register-module')(module)
 
 const {getUploadControls} = require('~/fb-runner-node/page/utils/utils-uploads')
-const setUploadControlsMaxSize = require('./set-upload-controls-max-size')
 
 module.exports = async function setMultipartForm (pageInstance, userData) {
   const uploadControls = getUploadControls(pageInstance)
@@ -27,14 +26,6 @@ module.exports = async function setMultipartForm (pageInstance, userData) {
 
   if (encType) {
     pageInstance.encType = true
-
-    /**
-     * Filter upload controls
-     */
-    const uploadControls = getUploadControls(pageInstance)
-      .filter(({_type}) => _type === 'fileupload')
-
-    setUploadControlsMaxSize(uploadControls)
   }
 
   return pageInstance
