+++ lib/page/set-multipart-form/set-multipart-form.unit.spec.js
@@ -1,28 +1,12 @@
 const test = require('tape')
 
-const setUploadControlsMaxSize = require('./set-upload-controls-max-size')
 const setMultipartForm = require('./set-multipart-form')
 
 const getContentTypeStub = (type) => {
   return () => type
 }
 
-test('When setMultipartForm is passed a page containing no upload controls', async t => {
-  const pageInstance = await setMultipartForm({
-    _id: 'page.test',
-    _type: 'page.form',
-    components: [{
-      _type: 'text'
-    }]
-  }, {req: {}})
-
-  t.equal(pageInstance.encType, undefined, 'it should not set page’s encType to true')
-  t.equal(pageInstance.components[0].maxSize, undefined, 'it should not set component’s maxSize')
-
-  t.end()
-})
-
-test('When setMultipartForm is passed a page containing upload controls', async t => {
+test('The pageInstance has upload controls', async t => {
   const pageInstance = await setMultipartForm({
     _id: 'page.test',
     _type: 'page.form',
@@ -32,12 +16,25 @@ test('When setMultipartForm is passed a page containing upload controls', async 
   }, {req: {}})
 
   t.equal(pageInstance.encType, true, 'it should set page’s encType to true')
-  t.equal(pageInstance.components[0].validation.maxSize, setUploadControlsMaxSize.defaultMaxSize, 'it should set component’s maxSize')
 
   t.end()
 })
 
-test('When the method in use is POST and the content-type should be multipart/form-data and it is', async t => {
+test('The pageInstance does not have upload controls', async t => {
+  const pageInstance = await setMultipartForm({
+    _id: 'page.test',
+    _type: 'page.form',
+    components: [{
+      _type: 'text'
+    }]
+  }, {req: {}})
+
+  t.equal(pageInstance.encType, undefined, 'it should not set page’s encType to true')
+
+  t.end()
+})
+
+test('The method is POST and the content-type is multipart/form-data', async t => {
   await setMultipartForm({
     _id: 'page.test',
     _type: 'page.form',
@@ -56,7 +53,7 @@ test('When the method in use is POST and the content-type should be multipart/fo
   t.end()
 })
 
-test('When the method in use is POST and the content-type should be multipart/form-data but is not', async t => {
+test('The method is POST and the content-type is not multipart/form-data', async t => {
   try {
     t.throws(await setMultipartForm({
       _id: 'page.test',
@@ -77,7 +74,7 @@ test('When the method in use is POST and the content-type should be multipart/fo
   t.end()
 })
 
-test('When the method in use is POST and the content-type should be application/x-www-form-urlencoded and it is', async t => {
+test('The method is POST and the content-type is application/x-www-form-urlencoded', async t => {
   await setMultipartForm({
     _id: 'page.test',
     _type: 'page.form',
@@ -96,7 +93,7 @@ test('When the method in use is POST and the content-type should be application/
   t.end()
 })
 
-test('When the method in use is POST and the content-type should be multipart/form-data but is not', async t => {
+test('The method is POST and the content-type is not multipart/form-data', async t => {
   try {
     t.throws(await setMultipartForm({
       _id: 'page.test',
@@ -118,7 +115,7 @@ test('When the method in use is POST and the content-type should be multipart/fo
   t.end()
 })
 
-test('When the method in use is POST and there is no content-type', async t => {
+test('The method is POST and there is no content-type', async t => {
   try {
     t.throws(await setMultipartForm({
       _id: 'page.test',
